Dynamically detect interface without provisioning network
When we are not using a dedicated provisioning network, the static IP
manager will attempt to determine the interface we want to use for the
external network dynamically.

That interface will already have a lease on the network, but the
iteration of optional provisioning network for 4.6 still has a
dependency on a VIP for provisioning services. This may be later
removed, but not in the 4.6 cycle.